Add static types to Scroller component
