+++ src/components/common/Scroller/ScrollerContainer/styles.scss
@@ -1,5 +1,5 @@
 // Global styles
-@import "../../../styles/base/settings";
+@import "../../../../styles/base/settings";
 
 // Ended content wrapper
 .ended {
