--- src/components/common/Scroller/constants/defaultProps.js
@@ -1,12 +1,0 @@
-// Default props
-export default {
-  core: {
-    state: {
-      isError: null,
-      meta: {},
-      mode: null,
-      pagination: null,
-      query: null
-    }
-  }
-};
